ESQL: Merge `Configuration`
This merges esql-core's `Configuration` into esql's `EsqlConfiguration`
and then renamed esql's configuration to `Configuration`. Now there's
only one configuration! Not two layers, one that was copied from ql.
After this we can make changes to our configuration to better suite our
needs much more easilly.

It also centralizes a place to grab test configuration into
`ConfigurationTestUtils` in esql's test fixtures. Most of that was
scattered between a test fixture and the serialization tests.
