+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/optimizer/LocalLogicalOptimizerContext.java
@@ -7,7 +7,7 @@
 
 package org.elasticsearch.xpack.esql.optimizer;
 
-import org.elasticsearch.xpack.esql.session.EsqlConfiguration;
+import org.elasticsearch.xpack.esql.session.Configuration;
 import org.elasticsearch.xpack.esql.stats.SearchStats;
 
 import java.util.Objects;
@@ -15,7 +15,7 @@ import java.util.Objects;
 public final class LocalLogicalOptimizerContext extends LogicalOptimizerContext {
     private final SearchStats searchStats;
 
-    public LocalLogicalOptimizerContext(EsqlConfiguration configuration, SearchStats searchStats) {
+    public LocalLogicalOptimizerContext(Configuration configuration, SearchStats searchStats) {
         super(configuration);
         this.searchStats = searchStats;
     }
