+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/plugin/ComputeService.java
@@ -69,7 +69,7 @@ import org.elasticsearch.xpack.esql.plan.physical.PhysicalPlan;
 import org.elasticsearch.xpack.esql.planner.EsPhysicalOperationProviders;
 import org.elasticsearch.xpack.esql.planner.LocalExecutionPlanner;
 import org.elasticsearch.xpack.esql.planner.PlannerUtils;
-import org.elasticsearch.xpack.esql.session.EsqlConfiguration;
+import org.elasticsearch.xpack.esql.session.Configuration;
 import org.elasticsearch.xpack.esql.session.Result;
 
 import java.util.ArrayList;
@@ -131,7 +131,7 @@ public class ComputeService {
         String sessionId,
         CancellableTask rootTask,
         PhysicalPlan physicalPlan,
-        EsqlConfiguration configuration,
+        Configuration configuration,
         ActionListener<Result> listener
     ) {
         Tuple<PhysicalPlan, PhysicalPlan> coordinatorAndDataNodePlan = PlannerUtils.breakPlanBetweenCoordinatorAndDataNode(
@@ -263,7 +263,7 @@ public class ComputeService {
         String sessionId,
         String clusterAlias,
         CancellableTask parentTask,
-        EsqlConfiguration configuration,
+        Configuration configuration,
         PhysicalPlan dataNodePlan,
         Set<String> concreteIndices,
         String[] originalIndices,
@@ -324,7 +324,7 @@ public class ComputeService {
     private void startComputeOnRemoteClusters(
         String sessionId,
         CancellableTask rootTask,
-        EsqlConfiguration configuration,
+        Configuration configuration,
         PhysicalPlan plan,
         ExchangeSourceHandler exchangeSource,
         List<RemoteCluster> clusters,
@@ -433,7 +433,7 @@ public class ComputeService {
     private void acquireSearchContexts(
         String clusterAlias,
         List<ShardId> shardIds,
-        EsqlConfiguration configuration,
+        Configuration configuration,
         Map<Index, AliasFilter> aliasFilters,
         ActionListener<List<SearchContext>> listener
     ) {
@@ -611,7 +611,7 @@ public class ComputeService {
         }
 
         private void runBatch(int startBatchIndex) {
-            final EsqlConfiguration configuration = request.configuration();
+            final Configuration configuration = request.configuration();
             final String clusterAlias = request.clusterAlias();
             final var sessionId = request.sessionId();
             final int endBatchIndex = Math.min(startBatchIndex + maxConcurrentShards, request.shardIds().size());
@@ -791,7 +791,7 @@ public class ComputeService {
         String clusterAlias,
         String globalSessionId,
         CancellableTask parentTask,
-        EsqlConfiguration configuration,
+        Configuration configuration,
         ExchangeSinkExec plan,
         Set<String> concreteIndices,
         String[] originalIndices,
@@ -839,7 +839,7 @@ public class ComputeService {
         String sessionId,
         String clusterAlias,
         List<SearchContext> searchContexts,
-        EsqlConfiguration configuration,
+        Configuration configuration,
         ExchangeSourceHandler exchangeSource,
         ExchangeSinkHandler exchangeSink
     ) {
