+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/SerializationTestUtils.java
@@ -37,7 +37,7 @@ import org.elasticsearch.xpack.esql.io.stream.PlanStreamOutput;
 import org.elasticsearch.xpack.esql.plan.logical.LogicalPlan;
 import org.elasticsearch.xpack.esql.plan.physical.PhysicalPlan;
 import org.elasticsearch.xpack.esql.querydsl.query.SingleValueQuery;
-import org.elasticsearch.xpack.esql.session.EsqlConfiguration;
+import org.elasticsearch.xpack.esql.session.Configuration;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -52,7 +52,7 @@ public class SerializationTestUtils {
         assertSerialization(plan, EsqlTestUtils.TEST_CFG);
     }
 
-    public static void assertSerialization(PhysicalPlan plan, EsqlConfiguration configuration) {
+    public static void assertSerialization(PhysicalPlan plan, Configuration configuration) {
         var deserPlan = serializeDeserialize(
             plan,
             PlanStreamOutput::writePhysicalPlanNode,
@@ -71,7 +71,7 @@ public class SerializationTestUtils {
         assertSerialization(expression, EsqlTestUtils.TEST_CFG);
     }
 
-    public static void assertSerialization(Expression expression, EsqlConfiguration configuration) {
+    public static void assertSerialization(Expression expression, Configuration configuration) {
         Expression deserExpression = serializeDeserialize(
             expression,
             PlanStreamOutput::writeNamedWriteable,
@@ -85,7 +85,7 @@ public class SerializationTestUtils {
         return serializeDeserialize(orig, serializer, deserializer, EsqlTestUtils.TEST_CFG);
     }
 
-    public static <T> T serializeDeserialize(T orig, Serializer<T> serializer, Deserializer<T> deserializer, EsqlConfiguration config) {
+    public static <T> T serializeDeserialize(T orig, Serializer<T> serializer, Deserializer<T> deserializer, Configuration config) {
         try (BytesStreamOutput out = new BytesStreamOutput()) {
             PlanStreamOutput planStreamOutput = new PlanStreamOutput(out, planNameRegistry, config);
             serializer.write(planStreamOutput, orig);
