+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/expression/function/scalar/AbstractConfigurationFunctionTestCase.java
@@ -15,7 +15,7 @@ import org.elasticsearch.xpack.esql.core.util.StringUtils;
 import org.elasticsearch.xpack.esql.expression.function.AbstractScalarFunctionTestCase;
 import org.elasticsearch.xpack.esql.plugin.EsqlPlugin;
 import org.elasticsearch.xpack.esql.plugin.QueryPragmas;
-import org.elasticsearch.xpack.esql.session.EsqlConfiguration;
+import org.elasticsearch.xpack.esql.session.Configuration;
 
 import java.util.List;
 import java.util.Map;
@@ -23,16 +23,16 @@ import java.util.Map;
 import static org.elasticsearch.xpack.esql.SerializationTestUtils.assertSerialization;
 
 public abstract class AbstractConfigurationFunctionTestCase extends AbstractScalarFunctionTestCase {
-    protected abstract Expression buildWithConfiguration(Source source, List<Expression> args, EsqlConfiguration configuration);
+    protected abstract Expression buildWithConfiguration(Source source, List<Expression> args, Configuration configuration);
 
     @Override
     protected Expression build(Source source, List<Expression> args) {
         return buildWithConfiguration(source, args, EsqlTestUtils.TEST_CFG);
     }
 
-    static EsqlConfiguration randomConfiguration() {
+    static Configuration randomConfiguration() {
         // TODO: Randomize the query and maybe the pragmas.
-        return new EsqlConfiguration(
+        return new Configuration(
             randomZone(),
             randomLocale(random()),
             randomBoolean() ? null : randomAlphaOfLength(randomInt(64)),
@@ -47,12 +47,12 @@ public abstract class AbstractConfigurationFunctionTestCase extends AbstractScal
     }
 
     public void testSerializationWithConfiguration() {
-        EsqlConfiguration config = randomConfiguration();
+        Configuration config = randomConfiguration();
         Expression expr = buildWithConfiguration(testCase.getSource(), testCase.getDataAsFields(), config);
 
         assertSerialization(expr, config);
 
-        EsqlConfiguration differentConfig;
+        Configuration differentConfig;
         do {
             differentConfig = randomConfiguration();
         } while (config.equals(differentConfig));
