+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/optimizer/TestPlannerOptimizer.java
@@ -15,7 +15,7 @@ import org.elasticsearch.xpack.esql.plan.physical.EstimatesRowSize;
 import org.elasticsearch.xpack.esql.plan.physical.PhysicalPlan;
 import org.elasticsearch.xpack.esql.planner.Mapper;
 import org.elasticsearch.xpack.esql.planner.PlannerUtils;
-import org.elasticsearch.xpack.esql.session.EsqlConfiguration;
+import org.elasticsearch.xpack.esql.session.Configuration;
 import org.elasticsearch.xpack.esql.stats.SearchStats;
 
 public class TestPlannerOptimizer {
@@ -25,9 +25,9 @@ public class TestPlannerOptimizer {
     private final PhysicalPlanOptimizer physicalPlanOptimizer;
     private final EsqlFunctionRegistry functionRegistry;
     private final Mapper mapper;
-    private final EsqlConfiguration config;
+    private final Configuration config;
 
-    public TestPlannerOptimizer(EsqlConfiguration config, Analyzer analyzer) {
+    public TestPlannerOptimizer(Configuration config, Analyzer analyzer) {
         this.analyzer = analyzer;
         this.config = config;
